10/31/2024 12 40    fetching the record from the database

+++ app/src/main/java/com/example/recipefinder/database/MealPlan/MealPlan.kt
@@ -8,5 +8,5 @@ data class MealPlan(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val date: String,
     val mealTime: String,
-    val dishId: Long
+    val dishId: Int
 )
